+++ packages/client3/pages/artist/create-pool/create-pool-document-upload.tsx
@@ -11,8 +11,7 @@ import UploadPDF, {
   UploadedFileType,
 } from "@/components/upload-pdf/upload-pdf";
 import { toBase64 } from "@/lib/file-utils";
-
-import { IPool } from "./create-pool-form";
+import { IPool } from "@/types/pool";
 
 type Props = {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
