+++ packages/client3/types/pool/index.ts
@@ -1,0 +1,1 @@
+export * from "./pool.types";
